Fix signed comparator.
Turns out I mistranslated an old optimized Verilog comparator of mine --
back before 2021, in fact. This has caused certain signed-less-than
comparisons to misfire. Heavily dependent on operand ordering, this bug
escaped my tests and only became apparent when I tried to use GCC's
division routines in Dhrystone.

I've added specific tests for BLT that cover this, as well as an extract
of the signed division routine from libgcc. It's a slow test but it
provides a lot of assurance.