box: hacked-up counter "support" (read only)
Setting the `counters` parameter to `True` will enable the cycle and
instret counters. The `rdcycle` / `rdinstret` instructions will start
working.

That's it. This is super minimal counter support intended for use in
Dhrystone and the like, it is _not_ full CSR support -- doing full CSR
support on CPU16 showed me just how expensive it is.

This version costs 153 LCs right now, which is a 19% size increase to
the minimal SoC example, but is tolerable in the context of an SoC large
enough to run Dhrystone. True CSRs would be worse.

Current implementation weirdnesses include

- Top half of counters not implemented (the fact that the spec treats
  them as mandatory seems like a Big Computer People decision)
- Literally any SYSTEM instruction will be interpreted as a CSRR.

In exchange, the entire implementation with 2x 32-bit counters costs 153
LCs.