icestick-smallest: add cheat detection code
My Bluespec RISC-V implementations sometimes generated misleadingly
small circuits because Yosys, to its credit, noticed that certain bits
in the toy program in ROM were constant... and optimized out parts of
the CPU.

I've been carefully avoiding that here through manual checks, but why
not automate it. Now if you run the "smallest" script, which is a sign
that you're trying to generate numbers to brag about, it will verify
that every bit on the databus can take two values, preventing this class
of optimization from making things get weird.