boxcpu: fix carry propagation in second SW halfword
This was a fun one, because it only affected _negative_ immediate
displacements in word-size store instructions, and only in SoCs with
greater than 16 bits of implemented address space.

By failing to chain the carry (or, really, preserve it) for the second
bus cycle of a word store, we effectively subtracted 1 from the top 16
bits of the address if the displacement was negative.

As you might imagine this caused some problems.