dhrystone: fix integer overflows, make RUNS an arg
The report-printing code in dhrystone is _full_ of integer overflows,
many of them apparently introduced for PicoRV32's output code. A lot of
them are harmless (addition/subtraction that briefly goes negative
because int but winds up with the right answer). But there are a fair
number of them that are truncating (particularly multiply-divide pairs).
The net effect of this was, if the cycle count or instructions-retired
count got high enough, the report printing would totally break.

Fixed that the clumsy way by converting all of those operations to
64-bit. This means we're now linking in 64-bit arithmetic routines, and
the binary is bigger.

But it also means I can increase the run count, so now the number of
runs is a parameter to the makefile instead of a hardcoded constant.