New optimization opportunity, now under 800 LCs.
I was squinting at the critical path in the adder and noticed this
opportunity, which has slightly improved speed, but also knocked off 16
LCs. Meaning, the faster version of the core is now under 800 LCs too.

I applied an equivalent optimization to chonk, in the interest of
maintaining the comparison between the cores, and it didn't help.