CPU: stop wild fetches during JAL/JALR
For simplicity reasons, the intermediate add result during JAL and JALR
was being presented to FD and sent out onto the bus as a fetch address.
This caused two semi-random memory accesses during execution of the
instruction, the results of which were thrown away as speculation
failures.

With this change, that still happens on the address lines, but the bus
valid signal is suppressed to prevent the reads from affecting anything.

For reasons I can't quite explain, this change makes the system slightly
smaller.